Move example activity configs into TestActivityFiles
